requirements.txt
datasets>=4.8.3
ipykernel>=7.2.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.10.8
python-dotenv>=1.2.2
langchain>=1.2.17
langchain-groq>=1.1.2
